Tweaked text in documentation slide to fit in the layout. Added help-browser icon which is displayed in line with text in documentation slide.
Changed welcome slide to mention the version ("Ubuntu 9.04"). This sounds a lot better to someone who already uses Ubuntu, but should make a little difference in tone for a user who _is_ new.
Changed middle line in welcome slide to describe the slideshow more precisely, rather than making possibly inaccurate claims. (Fixes bug: lp:426620)

Tweaked layout slightly to give more room to the left and top of the list.
"special" class is now designed for use outside of a list, so that a special item is not padded unnecessarily on the sides.
Added "img.inline" rule in general.css. These images are set to 14 pixels high to prevent clashes.