Major adjustment to format for passing parameters to slideshow's Javascript via #parameter. Now using #?key?key=value  format, for compatibility with WebKit. This change is due to an undocumented bug where webkit.WebView.open() would mangle the URI string, leaving only the first occurrence of  the # character intact.
This change does not break compatibility with existing Ubiquity code, since the first parameter will be read whether it has a ? before it or not. However, Ubiquity should be updated as the previous functionality did not allow for multiple parameters.

Continuing adjustments to Slideshow.py. It now accepts parameters over the command line with OptionParser. Parameters accepted are locale, rtl, and slideshow path.
SlideshowViewer object handles basic selection of the slideshow language, choosing the ideal one of the available translations.
Added some basic documentation for the SlideshowViewer widget.